also move trace_compile trace_dispatch macros
