+++ base/timing.jl
@@ -630,7 +630,7 @@ macro timed(ex)
 end
 
 # Exported, documented, and tested in InteractiveUtils
-# here so it's possible to time all imports, including InteractiveUtils and its deps
+# here so it's possible to time/trace all imports, including InteractiveUtils and its deps
 macro time_imports(ex)
     quote
         try
@@ -638,6 +638,28 @@ macro time_imports(ex)
             $(esc(ex))
         finally
             Base.Threads.atomic_sub!(Base.TIMING_IMPORTS, 1)
+        end
+    end
+end
+
+macro trace_compile(ex)
+    quote
+        try
+            ccall(:jl_force_trace_compile_timing_enable, Cvoid, ())
+            $(esc(ex))
+        finally
+            ccall(:jl_force_trace_compile_timing_disable, Cvoid, ())
+        end
+    end
+end
+
+macro trace_dispatch(ex)
+    quote
+        try
+            ccall(:jl_force_trace_dispatch_enable, Cvoid, ())
+            $(esc(ex))
+        finally
+            ccall(:jl_force_trace_dispatch_disable, Cvoid, ())
         end
     end
 end
