+++ stdlib/InteractiveUtils/src/macros.jl
@@ -5,8 +5,8 @@
 import Base: typesof, insert!, replace_ref_begin_end!, infer_effects
 
 # defined in Base so it's possible to time all imports, including InteractiveUtils and its deps
-# via. `Base.@time_imports`
-import Base: @time_imports
+# via. `Base.@time_imports` etc.
+import Base: @time_imports, @trace_compile, @trace_dispatch
 
 separate_kwargs(args...; kwargs...) = (args, values(kwargs))
 
@@ -246,28 +246,6 @@ macro code_lowered(ex0...)
     quote
         local results = $thecall
         length(results) == 1 ? results[1] : results
-    end
-end
-
-macro trace_compile(ex)
-    quote
-        try
-            ccall(:jl_force_trace_compile_timing_enable, Cvoid, ())
-            $(esc(ex))
-        finally
-            ccall(:jl_force_trace_compile_timing_disable, Cvoid, ())
-        end
-    end
-end
-
-macro trace_dispatch(ex)
-    quote
-        try
-            ccall(:jl_force_trace_dispatch_enable, Cvoid, ())
-            $(esc(ex))
-        finally
-            ccall(:jl_force_trace_dispatch_disable, Cvoid, ())
-        end
     end
 end
